+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## [2.4.1]
+- #125 Evol Apidae : ne plus utiliser prestations.labelsTourismeHandicap
 - EVO - Amélioration de la configuration entre env de production / cooking [#116](https://github.com/GeotrekCE/Sitourisme/issues/116)
 - #121 - Synchroniser les latitudes et longitudes sur Apidae Cooking
+- Fix conf, Gpx, geoloc, console
 
 ## [2.4.0](https://github.com/GeotrekCE/Sitourisme/releases/tag/2.4.0) - December 2025
 - EVO 5.4 - Uploader les GPX sur Apidae [#49](https://github.com/GeotrekCE/Sitourisme/issues/49)